# doc/api.yaml
openapi: 3.0.0
paths:
...
  /auth/signup:
    post:
      operationId: AuthController_signup
      summary: Signup
      description: Signup a user
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: The user has been successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Request body does not contain required fields
      tags: &ref_10
        - auth
  /auth/login:
    post:
      operationId: AuthController_login
      summary: Login
      description: Logins a user and returns access & refresh JWT tokens
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '200':
          description: Successful login with provided login and password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDto'
        '400':
          description: DTO is invalid (no login or password, or they are not a strings)
        '403':
          description: Incorrect login or password
      tags: *ref_10
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      summary: Refresh
      description: Refreshes user's JWT tokens
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenDto'
      responses:
        '200':
          description: Successful refresh, new tokens issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponseDto'
        '401':
          description: DTO is invalid (no refreshToken in body)
        '403':
          description: >-
            Authentication failed (refresh token in the body is invalid or
            expired)
      tags: *ref_10
info:
  title: Home Library Service
  description: 'Home Library Service: RS School NodeJS 2023 Q2 - Weeks 6-9 Task'
  version: 0.0.1
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          description: user ID (uuid v4)
          example: 83ef2c0c-967e-46bb-b748-6614bc99b7b4
        login:
          type: string
          description: user login
          example: johndoe
        version:
          type: number
          minimum: 1
          description: user version (integer number, increments on update)
          example: 1
        createdAt:
          type: number
          description: timestamp of user creation
          example: 1690366188813
        updatedAt:
          type: number
          description: timestamp of last user update
          example: 1690366221355
      required:
        - id
        - login
        - version
        - createdAt
        - updatedAt
    CreateUserDto:
      type: object
      properties:
        login:
          type: object
          default: ''
          description: user login
          example: johndoe
        password:
          type: object
          default: ''
          description: user password
          example: abracadabra
      required:
        - login
        - password
...
    TokenResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT access token
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzZmU4Y2VkNS1iNjc2LTQ1YTQtYWI5MC05ZGMwZjcyYjg0N2IiLCJpYXQiOjE2OTIxODI2MDUsImV4cCI6MTY5MjE4NjIwNX0.MvEHTBg1y1Vak4chpltrsxwqK7oCQcZYELD6EinIxWo
        refreshToken:
          type: string
          description: JWT refresh token
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIzZmU4Y2VkNS1iNjc2LTQ1YTQtYWI5MC05ZGMwZjcyYjg0N2IiLCJpYXQiOjE2OTIxODI2MDUsImV4cCI6MTY5MjI2OTAwNX0.LUgxkI_Yd1UggbCKiW8MoZu2G8k-376eR5Qasg7roSg
      required:
        - accessToken
        - refreshToken
    RefreshTokenDto:
      type: object
      properties:
        refreshToken:
          type: string
          description: JWT refresh token
          example: >-
            eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJjNmRjOWE1Zi1mMGM2LTQ2YzctODZhZC0yMTQwYjQzZDViN2IiLCJpYXQiOjE2OTIxOTkwMzcsImV4cCI6MTY5MjI4NTQzN30.bwlNClmODzRgLJcCSP0Zr4pjYlfHMgKIK5o7C_9CUeg
      required:
        - refreshToken